reorganized all args into one file
