+++ nail/src/pipeline/align_step.rs
@@ -7,7 +7,7 @@ use libnail::{
     structs::{Profile, Sequence},
 };
 
-use super::AlignArgs;
+use crate::args::AlignArgs;
 
 pub trait AlignStep: dyn_clone::DynClone {
     fn run(
